# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Unified observability platform for AI agents, APIs, and infrastructure.",
        "title": "Keelwave API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/v1",
    "paths": {
...
        "/admin/orgs/{orgID}/members": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/members"
                ],
                "summary": "List org members",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization UUID",
                        "name": "orgID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.MemberWithUser"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    }
                }
            }
        },
        "/admin/orgs/{orgID}/members/{userID}": {
            "delete": {
                "description": "Admin+. Cannot remove an owner or yourself.",
                "tags": [
                    "admin/members"
                ],
                "summary": "Remove a member from the org",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization UUID",
                        "name": "orgID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Member user UUID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        },
        "/admin/orgs/{orgID}/members/{userID}/role": {
            "patch": {
                "description": "Owner only. Cannot change an owner's role or promote to owner.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/members"
                ],
                "summary": "Update a member's role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization UUID",
                        "name": "orgID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Member user UUID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.updateRolePayload"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        },
...
        "/auth/oauth/{provider}/callback": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "OAuth provider callback — issues session and redirects to dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth provider (google, github)",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CSRF state",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Temporary Redirect"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/auth/oauth/{provider}/start": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Initiate OAuth login flow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth provider (google, github)",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Post-login redirect path",
                        "name": "redirect",
                        "in": "query"
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Temporary Redirect"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.updateRolePayload": {
            "type": "object",
            "required": [
                "role"
            ],
            "properties": {
                "role": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "member"
                    ]
                }
            }
        },
...
        "store.MemberWithUser": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Format: \"Bearer kw_\u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}